Update to Fedora 43

+++ Containerfile
@@ -1,12 +1,11 @@
-FROM quay.io/fedora/fedora:42
+FROM quay.io/fedora/fedora:43
 
 RUN dnf install -y glibc-langpack-en util-linux mesa-{libGL,libEGL,vulkan-drivers}.i686 \
                    mangohud.{x86_64,i686} gamescope                                     \
                    foot fzf icoextract ImageMagick jq                                   \
                    egl-utils vulkan-tools jstest-gtk
-
 RUN <<-EOF
-	dnf install -y https://github.com/Open-Wine-Components/umu-launcher/releases/latest/download/umu-launcher-1.2.9.fc42.rpm
+	dnf install -y https://github.com/Open-Wine-Components/umu-launcher/releases/latest/download/umu-launcher-1.3.0.fc43.rpm
 
 	cat <<-CAT >/usr/bin/umu-run
 		#!/usr/bin/python3
@@ -31,6 +30,11 @@ RUN <<-EOF
 	PYTHON
 
 	rm -r /root/.cache
+
+	if ! [ -e $XDG_DATA_HOME/Steam/compatibilitytools.d/* ]
+	then
+	    exit 1
+	fi
 	ln -s $XDG_DATA_HOME/Steam/compatibilitytools.d/{*,Proton}
 
 	chown -R root:root $XDG_DATA_HOME
